@@ -1,12 +1,11 @@
 import { Logger, NoopLogger } from '@dandi/core'
+import { MarbleValues } from '@rxjs-stuff/marbles'
 import { AnyFn, impersonate, stubLoggerFactory } from '@ts-mc/common/testing'
 import { RequestClientFixture, requestClientFixture } from '@ts-mc/core/client/testing'
 import { ServerEvent } from '@ts-mc/core/server-events'
 import { PlayerWithHeldItems } from '@ts-mc/core/types'
 import { MinigameEvents } from '@ts-mc/minigames'
-import { MarbleKey, MarbleValues } from '@rxjs-marbles'
 import { Observable } from 'rxjs'
-import { map, share } from 'rxjs/operators'
 
 export type PlayerMarbleValues = MarbleValues<PlayerWithHeldItems>
 export type ServerEventMarbleValues = MarbleValues<ServerEvent>
@@ -14,7 +13,7 @@ export type ServerEventMarbleValues = MarbleValues<ServerEvent>
 export interface MinigameEventsHelpers<TEvents extends MinigameEvents> {
   playerValues: PlayerMarbleValues
   serverEvents: ServerEventMarbleValues
-  initEvents(source$: Observable<MarbleKey>): TEvents
+  initEvents(source$: Observable<ServerEvent>): TEvents
 }
 
 export interface MinigameEventsFactory<TEvents extends MinigameEvents> {
@@ -40,14 +39,9 @@ export class MinigameEventsHelpersImpl<TEvents extends MinigameEvents> implement
 
   constructor(private readonly minigameEventsFactory: MinigameEventsFactory<TEvents>) {}
 
-  public initEvents(source$: Observable<MarbleKey>): TEvents {
-    const serverEvents$ = source$.pipe(
-      map(key => this.serverEvents[key]),
-      share(),
-    )
-
+  public initEvents(source$: Observable<ServerEvent>): TEvents {
     const playerValues = this.playerValues
-    const events = this.minigameEventsFactory(requestClientFixture(), serverEvents$, new NoopLogger())
+    const events = this.minigameEventsFactory(requestClientFixture(), source$, new NoopLogger())
 
     impersonate(events, function(this: MinigameEvents) {
       Object.values(playerValues || {}).forEach(player => this.addPlayer(player))
